requirements.txt
azureml-core
click
requests
